requirements.txt
deskew==1.5.1
huggingface_hub==0.24.6
safetensors==0.4.5
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==10.4.0
scikit-image==0.21.0
scipy==1.9.3
torch==2.2.1
torchvision==0.17.1
